requirements.txt
psycopg2
flask
flask_session
flask_login
pre-commit
python-dotenv